backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.3.1
qdrant-client==1.8.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.0